refactor: segment root runtime facades by domain
