+++ rustfs/src/server/http.rs
@@ -30,11 +30,11 @@ use crate::server::{
         TlsAcceptorHolder, TlsHandshakeFailureKind, build_acceptor_from_loaded, load_tls_material, spawn_reload_loop,
     },
 };
-use crate::storage_api::server as storage;
-use crate::storage_api::server::request_context::{RequestContext, extract_request_id_from_headers};
-use crate::storage_api::server::rpc::InternodeRpcService;
-use crate::storage_api::server::tonic_service::make_server;
-use crate::storage_api::server::{TONIC_RPC_PREFIX, verify_rpc_signature};
+use crate::storage_api::server::http as storage;
+use crate::storage_api::server::http::request_context::{RequestContext, extract_request_id_from_headers};
+use crate::storage_api::server::http::rpc::InternodeRpcService;
+use crate::storage_api::server::http::tonic_service::make_server;
+use crate::storage_api::server::http::{TONIC_RPC_PREFIX, verify_rpc_signature};
 use bytes::Bytes;
 use http::{HeaderMap, Method, Request as HttpRequest, Response};
 use hyper::body::Incoming;
@@ -1104,7 +1104,8 @@ fn process_connection(
                 .layer(
                     TraceLayer::new_for_http()
                         .make_span_with(|request: &HttpRequest<_>| {
-                            let request_context = request.extensions().get::<crate::storage_api::server::request_context::RequestContext>();
+                            let request_context =
+                                request.extensions().get::<crate::storage_api::server::http::request_context::RequestContext>();
                             let request_id = request_context
                                 .map(|ctx| ctx.request_id.as_str())
                                 .unwrap_or("unknown");
@@ -1253,7 +1254,8 @@ fn process_connection(
                 .layer(
                     TraceLayer::new_for_http()
                         .make_span_with(|request: &HttpRequest<_>| {
-                            let request_context = request.extensions().get::<crate::storage_api::server::request_context::RequestContext>();
+                            let request_context =
+                                request.extensions().get::<crate::storage_api::server::http::request_context::RequestContext>();
                             let request_id = request_context
                                 .map(|ctx| ctx.request_id.as_str())
                                 .unwrap_or("unknown");
