+++ rustfs/src/startup_bucket_metadata.rs
@@ -13,8 +13,10 @@
 // limitations under the License.
 
 use crate::startup_runtime_sources;
-use crate::storage_api::startup::contract::bucket::{BucketOperations, BucketOptions};
-use crate::storage_api::startup::{ECStore, init_bucket_metadata_sys, try_migrate_bucket_metadata, try_migrate_iam_config};
+use crate::storage_api::startup::bucket_metadata::contract::bucket::{BucketOperations, BucketOptions};
+use crate::storage_api::startup::bucket_metadata::{
+    ECStore, init_bucket_metadata_sys, try_migrate_bucket_metadata, try_migrate_iam_config,
+};
 use std::{
     io::{Error, Result},
     sync::Arc,
