+++ rustfs/src/startup_runtime_sources.rs
@@ -14,7 +14,7 @@
 
 use crate::config::RustFSBufferConfig;
 use crate::runtime_sources::{resolve_outbound_tls_generation, resolve_replication_pool_handle};
-use crate::storage_api::startup::{DynReplicationPool, set_global_region, set_global_rustfs_port};
+use crate::storage_api::startup::runtime_sources::{DynReplicationPool, set_global_region, set_global_rustfs_port};
 use rustfs_kms::KmsServiceManager;
 use rustfs_obs::{GlobalError as ObservabilityError, OtelGuard};
 use rustfs_tls_runtime::{OutboundTlsMaterial, TlsGeneration};
