+++ rustfs/src/storage_api.rs
@@ -18,7 +18,9 @@
 use rustfs_storage_api as storage_contracts;
 
 pub(crate) mod capacity {
-    pub(crate) use crate::storage::{all_local_disk, disk_drive_path, disk_endpoint};
+    pub(crate) mod service {
+        pub(crate) use crate::storage::{all_local_disk, disk_drive_path, disk_endpoint};
+    }
 }
 
 pub(crate) mod cluster {
@@ -71,87 +73,158 @@ pub(crate) mod error {
 }
 
 pub(crate) mod protocols {
-    pub(crate) mod access {
+    pub(crate) mod client {
         pub(crate) use crate::storage::access::ReqInfo;
-    }
-
-    pub(crate) mod ecfs {
         pub(crate) type FS = crate::storage::FS;
-    }
-
-    pub(crate) mod request_context {
         pub(crate) use crate::storage::request_context::RequestContext;
     }
 }
 
 pub(crate) mod server {
-    pub(crate) mod contract {
-        pub(crate) mod admin {
-            pub(crate) use super::super::super::storage_contracts::StorageAdminApi;
+    pub(crate) mod event {
+        pub(crate) mod contract {
+            #[cfg(test)]
+            pub(crate) mod lifecycle {
+                pub(crate) use super::super::super::super::storage_contracts::TransitionedObject;
+            }
         }
+
+        pub(crate) use crate::storage::{EventArgs, StorageObjectInfo, register_event_dispatch_hook};
+    }
+
+    pub(crate) mod http {
+        pub(crate) use crate::storage::{TONIC_RPC_PREFIX, verify_rpc_signature};
+
+        pub(crate) mod ecfs {
+            pub(crate) type FS = crate::storage::FS;
+        }
+
+        pub(crate) mod request_context {
+            pub(crate) use crate::storage::request_context::{RequestContext, extract_request_id_from_headers};
+        }
+
+        pub(crate) mod rpc {
+            pub(crate) use crate::storage::rpc::InternodeRpcService;
+        }
+
+        pub(crate) mod tonic_service {
+            pub(crate) use crate::storage::tonic_service::make_server;
+        }
+    }
+
+    pub(crate) mod layer {
+        pub(crate) use crate::storage::apply_cors_headers;
+
+        pub(crate) mod request_context {
+            pub(crate) use crate::storage::request_context::{
+                RequestContext, extract_request_id_from_headers, extract_trace_context_ids_from_headers, spawn_traced,
+            };
+        }
+    }
+
+    pub(crate) mod module_switch {
+        pub(crate) use crate::storage::{Error, read_config, save_config};
+    }
+
+    pub(crate) mod readiness {
+        pub(crate) mod contract {
+            pub(crate) mod admin {
+                pub(crate) use super::super::super::super::storage_contracts::StorageAdminApi;
+            }
+        }
+
+        pub(crate) use crate::storage::{Endpoint, EndpointServerPools, is_dist_erasure};
 
         #[cfg(test)]
-        pub(crate) mod lifecycle {
-            pub(crate) use super::super::super::storage_contracts::TransitionedObject;
-        }
+        pub(crate) use crate::storage::{Endpoints, PoolEndpoints};
     }
 
-    pub(crate) use crate::storage::{
-        ECStore, Endpoint, EndpointServerPools, Error, EventArgs, StorageObjectInfo, TONIC_RPC_PREFIX, apply_cors_headers,
-        is_dist_erasure, read_config, register_event_dispatch_hook, save_config, verify_rpc_signature,
-    };
-
-    #[cfg(test)]
-    pub(crate) use crate::storage::{Endpoints, PoolEndpoints};
-
-    pub(crate) mod ecfs {
-        pub(crate) type FS = crate::storage::FS;
-    }
-
-    pub(crate) mod request_context {
-        pub(crate) use crate::storage::request_context::{
-            RequestContext, extract_request_id_from_headers, extract_trace_context_ids_from_headers, spawn_traced,
-        };
-    }
-
-    pub(crate) mod rpc {
-        pub(crate) use crate::storage::rpc::InternodeRpcService;
-    }
-
-    pub(crate) mod tonic_service {
-        pub(crate) use crate::storage::tonic_service::make_server;
+    pub(crate) mod runtime_sources {
+        pub(crate) use crate::storage::{ECStore, EndpointServerPools};
     }
 }
 
 pub(crate) mod startup {
-    pub(crate) mod contract {
-        pub(crate) mod bucket {
-            pub(crate) use super::super::super::storage_contracts::{BucketOperations, BucketOptions};
+    pub(crate) mod background {
+        pub(crate) use crate::storage::{ECStore, set_workload_admission_snapshot_provider};
+    }
+
+    pub(crate) mod bucket_metadata {
+        pub(crate) mod contract {
+            pub(crate) mod bucket {
+                pub(crate) use super::super::super::super::storage_contracts::{BucketOperations, BucketOptions};
+            }
         }
+
+        pub(crate) use crate::storage::{ECStore, init_bucket_metadata_sys, try_migrate_bucket_metadata, try_migrate_iam_config};
     }
 
-    pub(crate) use crate::storage::{
-        DynReplicationPool, ECStore, EndpointServerPools, Result, get_bucket_notification_config, init_background_replication,
-        init_bucket_metadata_sys, init_ecstore_config, init_global_config_sys, init_local_disks, init_lock_clients,
-        new_global_notification_sys, prewarm_local_disk_id_map, process_lambda_configurations, process_queue_configurations,
-        process_topic_configurations, set_global_endpoints, set_global_region, set_global_rustfs_port,
-        set_workload_admission_snapshot_provider, shutdown_background_services, try_migrate_bucket_metadata,
-        try_migrate_iam_config, try_migrate_server_config, update_erasure_type,
-    };
-
-    #[cfg(test)]
-    pub(crate) use crate::storage::Error;
-
-    pub(crate) mod concurrency {
-        pub(crate) use crate::storage::concurrency::get_concurrency_manager;
-    }
-
-    pub(crate) mod deadlock_detector {
+    pub(crate) mod deadlock {
         pub(crate) use crate::storage::deadlock_detector::get_deadlock_detector;
     }
 
-    pub(crate) mod ecfs {
-        pub(crate) type FS = crate::storage::FS;
+    pub(crate) mod fs_guard {
+        pub(crate) use crate::storage::EndpointServerPools;
+    }
+
+    pub(crate) mod iam {
+        pub(crate) use crate::storage::ECStore;
+    }
+
+    pub(crate) mod init {
+        pub(crate) use crate::storage::{
+            get_bucket_notification_config, process_lambda_configurations, process_queue_configurations,
+            process_topic_configurations,
+        };
+
+        pub(crate) mod concurrency {
+            pub(crate) use crate::storage::concurrency::get_concurrency_manager;
+        }
+
+        pub(crate) mod ecfs {
+            pub(crate) type FS = crate::storage::FS;
+        }
+    }
+
+    pub(crate) mod lifecycle {
+        pub(crate) use crate::storage::ECStore;
+    }
+
+    pub(crate) mod notification {
+        pub(crate) use crate::storage::{EndpointServerPools, Result, new_global_notification_sys};
+
+        #[cfg(test)]
+        pub(crate) use crate::storage::Error;
+    }
+
+    pub(crate) mod runtime_sources {
+        pub(crate) use crate::storage::{DynReplicationPool, set_global_region, set_global_rustfs_port};
+    }
+
+    pub(crate) mod services {
+        pub(crate) use crate::storage::{ECStore, EndpointServerPools};
+    }
+
+    pub(crate) mod shutdown {
+        pub(crate) use crate::storage::shutdown_background_services;
+    }
+
+    pub(crate) mod storage {
+        pub(crate) use crate::storage::{
+            ECStore, EndpointServerPools, init_background_replication, init_ecstore_config, init_global_config_sys,
+            init_local_disks, init_lock_clients, prewarm_local_disk_id_map, set_global_endpoints, try_migrate_server_config,
+            update_erasure_type,
+        };
+    }
+}
+
+pub(crate) mod workload {
+    pub(crate) mod admission {
+        pub(crate) use crate::storage::{bucket_metadata_runtime_initialized, replication_queue_current_count};
+    }
+
+    pub(crate) mod concurrency {
+        pub(crate) use crate::storage::concurrency::get_concurrency_manager;
     }
 }
 
@@ -186,12 +259,4 @@ pub(crate) mod table {
         StorageObjectInfo, StorageObjectOptions, StorageObjectToDelete, StoragePutObjReader, get_bucket_metadata,
         get_lock_acquire_timeout, table_catalog_path_hash,
     };
-}
-
-pub(crate) mod workload {
-    pub(crate) use crate::storage::{bucket_metadata_runtime_initialized, replication_queue_current_count};
-
-    pub(crate) mod concurrency {
-        pub(crate) use crate::storage::concurrency::get_concurrency_manager;
-    }
 }
